backend/requirements.txt
# FastAPI & Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy[asyncio]>=2.0.25
aiosqlite>=0.19.0
asyncpg>=0.29.0
alembic>=1.13.0

# Binance
python-binance>=1.0.19

# Anthropic
anthropic>=0.18.0

# Charts
mplfinance>=0.12.10b0
matplotlib>=3.8.0
pandas>=2.1.0
numpy>=1.26.0

# Scheduler
apscheduler>=3.10.4
pytz>=2024.1

# Alerts
python-telegram-bot>=21.0
aiosmtplib>=3.0.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1  # Pin to 4.0.1 for passlib compatibility

# Utils
httpx>=0.26.0
python-dotenv>=1.0.0
aiofiles>=23.2.1

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
